+++ .github/workflows/automated_tests.yml
@@ -24,9 +24,9 @@ jobs:
     steps:
       - name: Checkout code
         uses: actions/checkout@v3.5.3
-      
-      - name: Run Docker 
+
+      - name: Run Docker
         run: docker-compose -f docker-compose.test.yml up --build -d
 
       - name: Test backend
-        run: cd mongo-backend && yarn install && yarn test        
+        run: yarn test:api
