+++ .gitignore
@@ -5,5 +5,6 @@ node_modules
 mongo-volume
 
 mongo-backend/.nyc_output
+mongo-backend/coverage
 
 yarn-error.log
