+++ mongo-backend/test/services/passport.test.js
@@ -1,6 +1,6 @@
 const bcrypt = require('bcrypt');
 const passport = require('passport');
-const User = require('../models/UserModel');
+const User = require('../../models/UserModel');
 const LocalStrategy = require('passport-local').Strategy;
 
 passport.serializeUser((user, done) => {
