--- mongo-backend/test/shipmentServices.test.js
@@ -1,13 +1,0 @@
-const shipmentServices = {
-  async getShipments() {},
-
-  async getShipmentById() {},
-
-  async createShipment() {},
-
-  async updateShipment() {},
-
-  async deleteShipment() {},
-};
-
-module.exports = shipmentServices;
